requirements.txt
Flask>=2.0.0
gunicorn>=20.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
